pyproject.toml
[tool.poetry]
name = "images"
version = "0.1.0"
description = "image retrieval"
authors = ["okahwaji <26441022+omarkahwaji@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.0.3"
numpy = "^1.25.2"
fastapi = "^0.101.0"
pydantic = "^2.1.1"
sqlalchemy = "^2.0.19"
python-dotenv = "^1.0.0"
opencv-python = "^4.8.0.74"
pillow = "^10.0.0"
sqlalchemy-utils = "^0.41.1"
uvicorn = "^0.23.2"
mysql-connector-python = "^8.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
flake8 = "^6.1.0"
tasks = "^2.8.0"
mypy = "^1.4.1"
pylint = "^2.17.5"
jupyter = "^1.0.0"
invoke = "^2.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
